Conversion corrections. Decreased 65536 to 65535 for SCTP server buffers.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.09.07@3185 0a995953-517f-4c62-b348-104708d926b2

+++ Kernel/spot-on-kernel.cc
@@ -3551,8 +3551,10 @@ void spoton_kernel::updateStatistics(void)
 	query.bindValue
 	  (0,
 	   QString::
-	   number(100 * static_cast<double> (v1) / v3 +
-		  100 * static_cast<double> (v2) / v3, 'f', 2).
+	   number(100.0 * static_cast<double> (v1) /
+		  static_cast<double> (v3) +
+		  100.0 * static_cast<double> (v2) /
+		  static_cast<double> (v3), 'f', 2).
 	   append("%"));
 	query.exec();
 	query.prepare("INSERT OR REPLACE INTO kernel_statistics "
